mailer_project/requirements.txt
PySocks==1.7.1
tkinterdnd2==0.3.0
pillow==10.4.0
requests==2.32.3